Refactor persistence namespaces and add ResourceRepository
This commit removes the `DigitalRightsManagement.Infrastructure.Persistence.Repositories` namespace and replaces it with `DigitalRightsManagement.Infrastructure.Persistence` across multiple files. This restructuring aims to simplify the organization of the persistence layer.

Affected files include `InfrastructureDependencyInjection.cs`, `ResourceOwnerAuthorizationBehavior.cs`, `TransactionBehavior.cs`, `PublisherExtensions.cs`, `ApplicationDbContext.cs`, `ApplicationDbManager.cs`, and several migration files.

Additionally, a new `ResourceRepository` class has been introduced in `ResourceRepository.cs`, which is currently empty but indicates future resource management functionality. These changes enhance maintainability and clarity in the project's structure.
