+++ src/DigitalRightsManagement.Infrastructure/Persistence/Repositories/ResourceRepository.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DigitalRightsManagement.Infrastructure.Persistence.Repositories
+{
+    class ResourceRepository
+    {
+    }
+}
